requirements.txt
Django==4.2.7
Pillow==10.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
channels==4.0.0
channels-redis==4.1.0
